加入異常處理機制

+++ src/main/java/idv/rennnhong/backendstarterkit/service/Impl/RoleServiceImpl.java
@@ -1,5 +1,6 @@
 package idv.rennnhong.backendstarterkit.service.Impl;
 
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
@@ -8,14 +9,16 @@ import idv.rennnhong.backendstarterkit.controller.request.role.RolePermissionDto
 import idv.rennnhong.backendstarterkit.controller.request.role.UpdateRoleRequestDto;
 import idv.rennnhong.backendstarterkit.dto.RoleDto;
 import idv.rennnhong.backendstarterkit.dto.mapper.RoleMapper;
-import idv.rennnhong.backendstarterkit.repository.ActionRepository;
-import idv.rennnhong.backendstarterkit.repository.PermissionRepository;
-import idv.rennnhong.backendstarterkit.repository.RoleRepository;
-import idv.rennnhong.backendstarterkit.repository.UserRepository;
+import idv.rennnhong.backendstarterkit.exception.ExceptionFactory;
+import idv.rennnhong.backendstarterkit.exception.ExceptionType;
 import idv.rennnhong.backendstarterkit.model.entity.Permission;
 import idv.rennnhong.backendstarterkit.model.entity.Role;
 import idv.rennnhong.backendstarterkit.model.entity.RolePermission;
 import idv.rennnhong.backendstarterkit.model.entity.User;
+import idv.rennnhong.backendstarterkit.repository.ActionRepository;
+import idv.rennnhong.backendstarterkit.repository.PermissionRepository;
+import idv.rennnhong.backendstarterkit.repository.RoleRepository;
+import idv.rennnhong.backendstarterkit.repository.UserRepository;
 import idv.rennnhong.backendstarterkit.service.RoleService;
 import idv.rennnhong.common.query.PageableResult;
 import idv.rennnhong.common.query.PageableResultImpl;
@@ -24,11 +27,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
-import java.util.UUID;
+import java.util.*;
+import java.util.stream.Collectors;
+
+import static idv.rennnhong.backendstarterkit.exception.GroupType.*;
 
 @Service
 public class RoleServiceImpl implements RoleService {
@@ -60,7 +64,6 @@ public class RoleServiceImpl implements RoleService {
 
     @Autowired
     public RoleServiceImpl(RoleRepository roleRepository, RoleMapper roleMapper) {
-//        super(roleRepository, roleMapper);
         this.roleRepository = roleRepository;
         this.roleMapper = roleMapper;
     }
@@ -74,7 +77,10 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public RoleDto getById(UUID id) {
-        Role role = roleRepository.findById(id).get();
+        Optional<Role> optionalRole = roleRepository.findById(id);
+        Role role = optionalRole.orElseThrow(() ->
+                ExceptionFactory.newException(ROLE, ExceptionType.ENTITY_NOT_FOUND, id.toString())
+        );
         return roleMapper.toDto(role);
     }
 
@@ -87,7 +93,11 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public RoleDto update(UUID id, UpdateRoleRequestDto updateRoleRequestDto) {
-        Role role = roleRepository.findById(id).get();
+        Optional<Role> optionalRole = roleRepository.findById(id);
+        Role role = optionalRole.orElseThrow(() ->
+                ExceptionFactory.newException(ROLE, ExceptionType.ENTITY_NOT_FOUND, id.toString())
+        );
+
         roleMapper.updateEntity(role, updateRoleRequestDto);
 
         Role updatedRole = roleRepository.save(role);
@@ -96,6 +106,18 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public void delete(UUID id) {
+        Optional<Role> optionalRole = roleRepository.findById(id);
+        Role role = optionalRole.orElseThrow(() ->
+                ExceptionFactory.newException(ROLE, ExceptionType.ENTITY_NOT_FOUND, id.toString())
+        );
+
+        /*
+        這邊Hibernate會自動用邏輯刪除的查詢方式查出users，delete=1
+        就不會被查出來，所以不用擔心因使用者實際還存在資料庫而刪不掉role的問題
+         */
+        if(role.getUsers().size() > 0)
+            throw ExceptionFactory.newException(ROLE, ExceptionType.ENTITY_EXIST_RELATED, id.toString());
+
         roleRepository.deleteById(id);
     }
 
@@ -105,15 +127,35 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public RoleDto updateRolePage(UUID roleId, UUID pageId, List<UUID> actionIds) {
-        Role role = roleRepository.findById(roleId).get();
-        Permission permission = permissionRepository.findById(pageId).get();
-        Set<RolePermission> rolePermissions = role.getRolePermissions();
-        rolePermissions.clear();
+    public RoleDto updateRolePermission(UUID roleId, UUID permissionId, List<UUID> actionIds) {
+        Optional<Role> optionalRole = roleRepository.findById(roleId);
+        Role role = optionalRole.orElseThrow(() ->
+                ExceptionFactory.newException(ROLE, ExceptionType.ENTITY_NOT_FOUND, roleId.toString())
+        );
 
+        Optional<Permission> optionalPermission = permissionRepository.findById(permissionId);
+        Permission permission = optionalPermission.orElseThrow(() ->
+                ExceptionFactory.newException(PERMISSION, ExceptionType.ENTITY_NOT_FOUND, permissionId.toString())
+        );
+
+        //取得該角色所有的RolePermission
+        Set<RolePermission> roleAllPermissions = role.getRolePermissions();
+
+        /*
+         篩選出role針對此permission的所有操作，若有n個操作則會有n筆資料
+         */
+        Set<RolePermission> targetRolePermissions = roleAllPermissions
+                .stream()
+                .filter(rolePermission -> Objects.equals(rolePermission.getPermission(), permission))
+                .collect(Collectors.toSet());
+
+        //先清掉對此permission的所有操作
+        roleAllPermissions.removeAll(targetRolePermissions);
+
+        //重新新增對此permission的操作
         actionIds.stream()
                 .forEach(actionId -> {
-                    rolePermissions
+                    roleAllPermissions
                             .add(new RolePermission(permission, actionRepository.findById(actionId).get()));
                 });
         Role updatedRole = roleRepository.save(role);
@@ -144,16 +186,27 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public Set<RoleDto> getRolesByUserId(UUID userId) {
-        User user = userRepository.findById(userId).get();
+        Optional<User> optionalUser = userRepository.findById(userId);
+        User user = optionalUser.orElseThrow(() ->
+                ExceptionFactory.newException(USER, ExceptionType.ENTITY_NOT_FOUND, userId.toString())
+        );
         Set<Role> roles = user.getRoles();
         return ImmutableSet.copyOf(roleMapper.toDto(roles));
     }
 
-    @Override
-    public boolean isRoleReferenced(UUID id) {
-        Role role = roleRepository.findById(id).get();
-        return role.getUsers().size() > 0;
-    }
+//    @Override
+//    //此role是否被任何user使用
+//    public boolean isRoleReferenced(UUID id) {
+//        Optional<Role> optionalRole = roleRepository.findById(id);
+//        Role role = optionalRole.orElseThrow(() ->
+//                ExceptionFactory.newException(ROLE, ExceptionType.ENTITY_NOT_FOUND, id.toString())
+//        );
+//
+//        if(role.getUsers().size() > 0)
+//
+//
+//        return role.getUsers().size() > 0;
+//    }
 
 //    @Override
 //    public List<ApiDTO> getAllApiByRole(RoleDTO role, String url, String httpMethod) {
